Refactor base classes; add message classes
BaseActor and ActorSystem were previously coupled with the Message class and contained a lot of assumptions regarding how the actors would communicate and behave. This commit is aimed to decouple BaseActor and ActorSystem from the Message class. The new MessageActor and MessageActorSystem are equivalent to the previous BaseActor and ActorSystem.